[trello-27]: Moved nocache decorator to userpanel.decorator, simplified add_customer_activity
